requirements.txt
scikit-learn
matplotlib
seaborn
numpy
pandas
streamlit
Pillow
